requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-jose[cryptography]
requests

langgraph
langchain-core
langchain-openai
langgraph-checkpoint-postgres
psycopg_pool
psycopg-binary

pydantic
supabase
tenacity